+++ packages/event-bus-client/src/plugin.ts
@@ -4,7 +4,6 @@ interface TanStackDevtoolsEvent<TEventName extends string, TPayload = any> {
   pluginId?: string // Optional pluginId to filter events by plugin
 }
 declare global {
-  // eslint-disable-next-line no-var
   var __TANSTACK_EVENT_TARGET__: EventTarget | null
 }
 
@@ -48,7 +47,7 @@ export class EventClient<
     )
     if (this.#retryCount < this.#maxRetries) {
       this.#retryCount++
-      this.#eventTarget().dispatchEvent(new CustomEvent('tanstack-connect'))
+      this.dispatchCustomEvent('tanstack-connect', {})
       return
     }
 
@@ -75,11 +74,6 @@ export class EventClient<
     this.#connected = false
     this.#connectIntervalId = null
     this.#connectEveryMs = 500
-
-    if (typeof CustomEvent !== 'undefined') {
-      this.#connectFunction()
-      this.startConnectLoop()
-    }
   }
 
   private startConnectLoop() {
@@ -151,11 +145,28 @@ export class EventClient<
     return this.#pluginId
   }
 
+  private dispatchCustomEventShim(eventName: string, detail: any) {
+    try {
+      const event = new Event(eventName, {
+        detail: detail,
+      } as any)
+      this.#eventTarget().dispatchEvent(event)
+    } catch (e) {
+      this.debugLog('Failed to dispatch shim event')
+    }
+  }
+
+  private dispatchCustomEvent(eventName: string, detail?: any) {
+    try {
+      this.#eventTarget().dispatchEvent(new CustomEvent(eventName, { detail }))
+    } catch (e) {
+      this.dispatchCustomEventShim(eventName, detail)
+    }
+  }
+
   private emitEventToBus(event: TanStackDevtoolsEvent<string, any>) {
     this.debugLog('Emitting event to client bus', event)
-    this.#eventTarget().dispatchEvent(
-      new CustomEvent('tanstack-dispatch-event', { detail: event }),
-    )
+    this.dispatchCustomEvent('tanstack-dispatch-event', event)
   }
 
   emit<
@@ -172,11 +183,17 @@ export class EventClient<
     // wait to connect to the bus
     if (!this.#connected) {
       this.debugLog('Bus not available, will be pushed as soon as connected')
-      return this.#queuedEvents.push({
+      this.#queuedEvents.push({
         type: `${this.#pluginId}:${eventSuffix}`,
         payload,
         pluginId: this.#pluginId,
       })
+      // start connection to event bus
+      if (typeof CustomEvent !== 'undefined') {
+        this.#connectFunction()
+        this.startConnectLoop()
+      }
+      return
     }
     // emit right now
     return this.emitEventToBus({